# project-root/api/openapi.yaml
openapi: 3.0.0
info:
  title: Clients BW --> CVLB
  description: Lista de Clients
  contact:
    email: ernaniwaldhelm@gmail.com
  version: 1.0.0
servers:
  - url: /
paths:
  /clientes:
    get:
      summary: Retorna uma lista de clientes
      description: Trás um vetor de objetos json contendo todos os dados dos clientes
      responses:
        "200":
          description: Um vetor json de objetos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
  /novoCliente:
    post:
      summary: Adiciona Cliente
      description: Adiciona um novo cliente à lista de clientes
      responses:
        "200":
          description: Adicionado à lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
        "401":
          description: "Cliente não enviado, erro no cadastro!"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /cliente/{docNumber}:
    put:
      summary: Edita um cliente pelo seu docNumber
      parameters:
        - name: docNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Documento JSON contendo os dados de um cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "401":
          description: "Cliente não editado, erro nos dados"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{docNumber}:
    get:
      summary: Retorna um cliente pelo seu docNumber
      parameters:
        - name: docNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Documento JSON contendo os dados de um cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "401":
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      summary: Remove um cliente pelo seu docNumber
      parameters:
        - name: docNumber
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "204":
          description: Cliente removido com sucesso
        "401":
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
components:
  schemas:
    Cliente:
      required:
        - docNumber
        - email
        - nome
      type: object
      properties:
        nome:
          type: string
          example: José
        sobrenome:
          type: string
          example: da Silva
        docNumber:
          type: integer
          example: 128
        email:
          type: string
          format: email
          example: silva@mail.com
        telefone:
          type: string
          example: "2299999999"
        nascimento:
          type: string
          format: date
          example: 2001-08-31